smartsales_uc1_uc4/requirements.txt
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0
python-dateutil==2.9.0
reportlab==4.2.0
twilio==8.10.2


python-dotenv==1.0.1


